+++ scripts/submit_bb.py
@@ -9,10 +9,10 @@ import qcore.constants as const
 import qcore.simulation_structure as sim_struct
 from qcore import utils, shared
 from qcore.config import host
+from qcore.qclogging import get_basic_logger
 from shared_workflow.load_config import load
 from shared_workflow.shared import set_wct, confirm, get_hf_nt
 from shared_workflow.shared_automated_workflow import submit_sl_script
-from shared_workflow.workflow_logger import get_basic_logger
 from shared_workflow.shared_template import write_sl_script
 
 default_wct = "00:30:00"