+++ scripts/submit_post_emod3d.py
@@ -5,14 +5,16 @@ import glob
 import argparse
 from logging import Logger
 
-import qcore.constants as const
-import qcore.simulation_structure as sim_struct
 from qcore import utils, binary_version
 from qcore.config import get_machine_config, host
+import qcore.constants as const
+from qcore.qclogging import get_basic_logger
+import qcore.simulation_structure as sim_struct
+
 from shared_workflow.shared import confirm
 from shared_workflow.shared_automated_workflow import submit_sl_script
 from shared_workflow.shared_template import write_sl_script
-from shared_workflow.workflow_logger import get_basic_logger
+
 
 merge_ts_name_prefix = "post_emod3d_merge_ts"